# Merge internal transformers into single traversal pass