--- a/src/babel/transformation/file/index.js
+++ b/src/babel/transformation/file/index.js
@@ -364,6 +364,7 @@ export default class File {
       ref._generated = true;
       ref.id = uid;
       ref.type = "FunctionDeclaration";
+      this.attachAuxiliaryComment(ref);
       this.path.unshiftContainer("body", ref);
     } else {
       ref._compact = true;