--- a/src/babel/transformation/helpers/build-binary-assignment-operator-transformer.js
+++ b/src/babel/transformation/helpers/build-binary-assignment-operator-transformer.js
@@ -10,10 +10,6 @@ export default function (exports, opts) {
     return t.assignmentExpression("=", left, right);
   };
 
-  exports.shouldVisit = function (node) {
-    return node.operator && (node.operator === opts.operator || node.operator === opts.operator + "=");
-  };
-
   exports.ExpressionStatement = function (node, parent, scope, file) {
     // hit the `AssignmentExpression` one below
     if (this.isCompletionRecord()) return;