--- a/src/babel/transformation/index.js
+++ b/src/babel/transformation/index.js
@@ -9,7 +9,7 @@ import transformers from "./transformers";
 for (var key in transformers) {
   var transformer = transformers[key];
   var metadata = transformer.metadata = transformer.metadata || {};
-  metadata.category = metadata.category || "builtin-basic";
+  metadata.group = metadata.group || "builtin-basic";
 }
 
 pipeline.addTransformers(transformers);