--- a/src/babel/transformation/templates/helper-ludicrous-in.js
+++ b/src/babel/transformation/templates/helper-ludicrous-in.js
@@ -1,1 +1,0 @@
-Object(RIGHT)[LEFT] !== undefined;