--- a/src/babel/transformation/templates/helper-ludicrous-proxy-create.js
+++ b/src/babel/transformation/templates/helper-ludicrous-proxy-create.js
@@ -1,4 +1,0 @@
-(function (proxy, directory) {
-  directory.push(proxy);
-  return proxy;
-})