--- a/src/babel/transformation/transformer.js
+++ b/src/babel/transformation/transformer.js
@@ -25,7 +25,6 @@ export default class Transformer {
     };
 
     this.manipulateOptions = take("manipulateOptions");
-    this.shouldVisit       = take("shouldVisit");
     this.metadata          = take("metadata") || {};
     this.parser            = take("parser");
     this.post              = take("post");
@@ -41,18 +40,6 @@ export default class Transformer {
 
     this.handlers = this.normalize(transformer);
     this.key      = transformerKey;
-
-    //
-
-    if (!this.shouldVisit && !this.handlers.enter && !this.handlers.exit) {
-      var types = Object.keys(this.handlers);
-      this.shouldVisit = function (node) {
-        for (var i = 0; i < types.length; i++) {
-          if (node.type === types[i]) return true;
-        }
-        return false;
-      };
-    }
   }
 
   normalize(transformer: Object): Object {