--- a/src/babel/transformation/transformers/es3/member-expression-literals.js
+++ b/src/babel/transformation/transformers/es3/member-expression-literals.js
@@ -1,10 +1,12 @@
 import * as t from "../../../types";
 
-export function MemberExpression(node) {
-  var prop = node.property;
-  if (!node.computed && t.isIdentifier(prop) && !t.isValidIdentifier(prop.name)) {
-    // foo.default -> foo["default"]
-    node.property = t.literal(prop.name);
-    node.computed = true;
+export var MemberExpression = {
+  exit(node) {
+    var prop = node.property;
+    if (!node.computed && t.isIdentifier(prop) && !t.isValidIdentifier(prop.name)) {
+      // foo.default -> foo["default"]
+      node.property = t.literal(prop.name);
+      node.computed = true;
+    }
   }
-}
+};