--- a/src/babel/transformation/transformers/es3/member-expression-literals.js
+++ b/src/babel/transformation/transformers/es3/member-expression-literals.js
@@ -1,5 +1,9 @@
 import * as t from "../../../types";
 
+export var metadata = {
+  category: "builtin-cleanup"
+};
+
 export var MemberExpression = {
   exit(node) {
     var prop = node.property;