--- a/src/babel/transformation/transformers/es3/member-expression-literals.js
+++ b/src/babel/transformation/transformers/es3/member-expression-literals.js
@@ -1,7 +1,7 @@
 import * as t from "../../../types";
 
 export var metadata = {
-  category: "builtin-trailing"
+  group: "builtin-trailing"
 };
 
 export var MemberExpression = {