--- a/src/babel/transformation/transformers/es3/property-literals.js
+++ b/src/babel/transformation/transformers/es3/property-literals.js
@@ -1,9 +1,11 @@
 import * as t from "../../../types";
 
-export function Property(node) {
-  var key = node.key;
-  if (!node.computed && t.isIdentifier(key) && !t.isValidIdentifier(key.name)) {
-    // default: "bar" -> "default": "bar"
-    node.key = t.literal(key.name);
+export var Property = {
+  exit(node) {
+    var key = node.key;
+    if (!node.computed && t.isIdentifier(key) && !t.isValidIdentifier(key.name)) {
+      // default: "bar" -> "default": "bar"
+      node.key = t.literal(key.name);
+    }
   }
-}
+};