--- a/src/babel/transformation/transformers/es3/property-literals.js
+++ b/src/babel/transformation/transformers/es3/property-literals.js
@@ -1,5 +1,9 @@
 import * as t from "../../../types";
 
+export var metadata = {
+  category: "builtin-cleanup"
+};
+
 export var Property = {
   exit(node) {
     var key = node.key;