--- a/src/babel/transformation/transformers/es6/arrow-functions.js
+++ b/src/babel/transformation/transformers/es6/arrow-functions.js
@@ -1,7 +1,5 @@
 import * as t from "../../../types";
 
-export var shouldVisit = t.isArrowFunctionExpression;
-
 export function ArrowFunctionExpression(node) {
   t.ensureBlock(node);