--- a/src/babel/transformation/transformers/es6/block-scoping.js
+++ b/src/babel/transformation/transformers/es6/block-scoping.js
@@ -37,10 +37,6 @@ function standardizeLets(declars) {
   }
 }
 
-export function shouldVisit(node) {
-  return t.isVariableDeclaration(node) && (node.kind === "let" || node.kind === "const");
-}
-
 export function VariableDeclaration(node, parent, scope, file) {
   if (!isLet(node, parent)) return;