--- a/src/babel/transformation/transformers/es6/classes.js
+++ b/src/babel/transformation/transformers/es6/classes.js
@@ -11,17 +11,17 @@ import * as t from "../../../types";
 
 const PROPERTY_COLLISION_METHOD_NAME = "__initializeProperties";
 
-export var shouldVisit = t.isClass;
-
 export function ClassDeclaration(node, parent, scope, file) {
   return t.variableDeclaration("let", [
     t.variableDeclarator(node.id, t.toExpression(node))
   ]);
 }
 
-export function ClassExpression(node, parent, scope, file) {
-  return new ClassTransformer(this, file).run();
-}
+export var ClassExpression = {
+  exit(node, parent, scope, file) {
+    return new ClassTransformer(this, file).run();
+  }
+};
 
 var collectPropertyReferencesVisitor = {
   Identifier: {