--- a/src/babel/transformation/transformers/es6/destructuring.js
+++ b/src/babel/transformation/transformers/es6/destructuring.js
@@ -1,6 +1,10 @@
 import * as messages from "../../../messages";
 import * as t from "../../../types";
 
+export var metadata = {
+  category: "builtin-advanced"
+};
+
 export function ForOfStatement(node, parent, scope, file) {
   var left = node.left;