--- a/src/babel/transformation/transformers/es6/modules.js
+++ b/src/babel/transformation/transformers/es6/modules.js
@@ -1,7 +1,5 @@
 import * as t from "../../../types";
 
-export { shouldVisit } from "../internal/modules";
-
 function keepBlockHoist(node, nodes) {
   if (node._blockHoist) {
     for (let i = 0; i < nodes.length; i++) {