--- a/src/babel/transformation/transformers/es6/modules.js
+++ b/src/babel/transformation/transformers/es6/modules.js
@@ -8,6 +8,10 @@ function keepBlockHoist(node, nodes) {
   }
 }
 
+export var metadata = {
+  category: "builtin-modules"
+};
+
 export function ImportDeclaration(node, parent, scope, file) {
   // flow type
   if (node.isType) return;