--- a/src/babel/transformation/transformers/es6/object-super.js
+++ b/src/babel/transformation/transformers/es6/object-super.js
@@ -1,8 +1,6 @@
 import ReplaceSupers from "../../helpers/replace-supers";
 import * as t from "../../../types";
 
-export var shouldVisit = t.isSuper;
-
 function Property(path, node, scope, getObjectRef, file) {
   if (!node.method) return;