--- a/src/babel/transformation/transformers/es6/regex.unicode.js
+++ b/src/babel/transformation/transformers/es6/regex.unicode.js
@@ -1,10 +1,6 @@
 import rewritePattern from "regexpu/rewrite-pattern";
 import * as regex from "../../helpers/regex";
 
-export function shouldVisit(node) {
-  return regex.is(node, "u");
-}
-
 export function Literal(node) {
   if (!regex.is(node, "u")) return;
   node.regex.pattern = rewritePattern(node.regex.pattern, node.regex.flags);