--- a/src/babel/transformation/transformers/es6/spread.js
+++ b/src/babel/transformation/transformers/es6/spread.js
@@ -44,8 +44,6 @@ function build(props, scope) {
   return nodes;
 }
 
-export var shouldVisit = t.isSpreadElement;
-
 export function ArrayExpression(node, parent, scope) {
   var elements = node.elements;
   if (!hasSpread(elements)) return;