--- a/src/babel/transformation/transformers/es7/async-functions.js
+++ b/src/babel/transformation/transformers/es7/async-functions.js
@@ -1,7 +1,3 @@
 export var metadata = {
   stage: 1
 };
-
-export function shouldVisit() {
-  return false;
-}