… src/babel/transformation/templates/helper-ludicrous-proxy-directory.js
… src/babel/transformation/transformers/es6/tail-call.js
… src/babel/transformation/transformers/es6/template-literals.js
… src/babel/transformation/transformers/es7/class-properties.js
… src/babel/transformation/transformers/es7/decorators.js
… src/babel/transformation/transformers/es7/do-expressions.js
… src/babel/transformation/transformers/es7/export-extensions.js
… src/babel/transformation/transformers/es7/object-rest-spread.js
… src/babel/transformation/transformers/es7/trailing-function-commas.js
… src/babel/transformation/transformers/index.js
… src/babel/transformation/transformers/internal/block-hoist.js
… src/babel/transformation/transformers/internal/hoist-directives.js
… src/babel/transformation/transformers/internal/module-formatter.js
… src/babel/transformation/transformers/internal/modules.js
… src/babel/transformation/transformers/internal/shadow-functions.js
… src/babel/transformation/transformers/internal/strict.js
… src/babel/transformation/transformers/internal/validation.js
… src/babel/transformation/transformers/minification/dead-code-elimination.js
… src/babel/transformation/transformers/minification/member-expression-literals.js
… src/babel/transformation/transformers/minification/property-literals.js
… src/babel/transformation/transformers/optimisation/flow.for-of.js
… src/babel/transformation/transformers/other/async-to-generator.js
… src/babel/transformation/transformers/other/bluebird-coroutines.js
… src/babel/transformation/transformers/other/flow.js
… src/babel/transformation/transformers/other/ludicrous.js
… src/babel/transformation/transformers/other/regenerator.js
… src/babel/transformation/transformers/other/runtime/index.js
… src/babel/transformation/transformers/other/strict.js
… src/babel/transformation/transformers/spec/block-scoped-functions.js
… src/babel/transformation/transformers/spec/function-name.js