Complicates the internal transformer state fairly signifcantly but adds a transformer metadata `category` property that allows the merging of transformer visitors with the same category.

This isn't ready to be merged yet. Currently around ~100 tests failing. These are a combination of mixed state issues and infinite recursion bugs due to the changes required in the traversal algorithm to accommodate multi-method visitors.

cc @amasad @DmitrySoshnikov