@sebmck 2015-05-07
**Before**

```
$ DEBUG=babel time babel tsserver.js --out-file tsserver.foo.js
  babel [BABEL] tsserver.js: Parse start +0ms
  babel [BABEL] tsserver.js: Parse stop +752ms
  babel [BABEL] tsserver.js: Start set AST +2ms
  babel [BABEL] tsserver.js: Start scope building +0ms
  babel [BABEL] tsserver.js: End scope building +3s
  babel [BABEL] tsserver.js: End set AST +893ms
  babel [BABEL] tsserver.js: Start prepass +0ms
  babel [BABEL] tsserver.js: End prepass +2s
  babel [BABEL] tsserver.js: Start module formatter init +0ms
  babel [BABEL] tsserver.js: End module formatter init +1s
  babel [BABEL] tsserver.js: Start transformer strict +0ms
  babel [BABEL] tsserver.js: Finish transformer strict +4ms
  babel [BABEL] tsserver.js: Start transformer _validation +0ms
  babel [BABEL] tsserver.js: Finish transformer _validation +698ms
  babel [BABEL] tsserver.js: Start transformer validation.react +0ms
  babel [BABEL] tsserver.js: Finish transformer validation.react +698ms
  babel [BABEL] tsserver.js: Start transformer spec.functionName +0ms
  babel [BABEL] tsserver.js: Finish transformer spec.functionName +708ms
  babel [BABEL] tsserver.js: Start transformer spec.blockScopedFunctions +0ms
  babel [BABEL] tsserver.js: Finish transformer spec.blockScopedFunctions +721ms
  babel [BABEL] tsserver.js: Start transformer es6.tailCall +0ms
  babel [BABEL] tsserver.js: Finish transformer es6.tailCall +1s
  babel [BABEL] tsserver.js: Start transformer _blockHoist +0ms
  babel [BABEL] tsserver.js: Finish transformer _blockHoist +699ms
  babel [BABEL] tsserver.js: Start transformer _shadowFunctions +0ms
  babel [BABEL] tsserver.js: Finish transformer _shadowFunctions +2s
  babel [BABEL] tsserver.js: Start transformer _strict +0ms
  babel [BABEL] tsserver.js: Finish transformer _strict +1ms
  babel [BABEL] tsserver.js: Start transformer _moduleFormatter +0ms
  babel [BABEL] tsserver.js: Finish transformer _moduleFormatter +0ms
  babel [BABEL] tsserver.js: Start transformer es3.propertyLiterals +0ms
  babel [BABEL] tsserver.js: Finish transformer es3.propertyLiterals +712ms
  babel [BABEL] tsserver.js: Start transformer es3.memberExpressionLiterals +0ms
  babel [BABEL] tsserver.js: Finish transformer es3.memberExpressionLiterals +744ms
  babel [BABEL] tsserver.js: Generation start +1ms
[BABEL] Note: The code generator has deoptimised the styling of "tsserver.js" as it exceeds the max of "100KB".
  babel [BABEL] tsserver.js: Generation end +2s
       18.10 real        18.34 user         0.27 sys
```

**After**

```
$ DEBUG=babel time babel tsserver.js --out-file tsserver.foo.js
  babel [BABEL] tsserver.js: Parse start +0ms
  babel [BABEL] tsserver.js: Parse stop +752ms
  babel [BABEL] tsserver.js: Start set AST +1ms
  babel [BABEL] tsserver.js: Start scope building +1ms
  babel [BABEL] tsserver.js: End scope building +3s
  babel [BABEL] tsserver.js: End set AST +904ms
  babel [BABEL] tsserver.js: Start module formatter init +0ms
  babel [BABEL] tsserver.js: End module formatter init +1s
  babel [BABEL] tsserver.js: Start transformer builtin +1ms
  babel [BABEL] tsserver.js: Finish transformer builtin +6s
  babel [BABEL] tsserver.js: Generation start +1ms
[BABEL] Note: The code generator has deoptimised the styling of "tsserver.js" as it exceeds the max of "100KB".
  babel [BABEL] tsserver.js: Generation end +2s
       15.07 real        15.32 user         0.27 sys
```

Not nearly as substantial as I'd hoped :disappointed: