@sebmck 2015-05-07
Actually, seems to make *some* difference. Compiling the entire Ember core source with Babel on `master` takes:

```
real	0m50.575s
user	0m51.257s
sys	0m0.478s
```

On this branch it takes:

```
real	0m35.426s
user	0m36.004s
sys	0m0.580s
```

Going to need to get some more numbers before I move forward with finishing this refactor though since I don't want to waste too much time if it's pointless.