@amasad 2015-05-07
:dancer: 
Trying to run it. Running into this issue:

```
class bar extends wow {
    constructor() {
        super();
        var x =() => this.wat();
    }
}
```

```
/Users/amasad/code/babel/lib/babel/transformation/file/index.js:616
      throw err;
            ^
TypeError: unknown: Cannot read property 'start' of undefined
    at TraversalPath.errorWithNode (/Users/amasad/code/babel/lib/babel/traversal/path/index.js:478:28)
    at TraversalPath.enter (/Users/amasad/code/babel/lib/babel/transformation/transformers/es6/classes.js:136:20)
    at TraversalPath.call (/Users/amasad/code/babel/lib/babel/traversal/path/index.js:750:28)
    at TraversalPath.visit (/Users/amasad/code/babel/lib/babel/traversal/path/index.js:781:10)
    at TraversalContext.visitSingle (/Users/amasad/code/babel/lib/babel/traversal/context.js:72:41)
    at TraversalContext.visit (/Users/amasad/code/babel/lib/babel/traversal/context.js:82:19)
    at Function.traverse.node (/Users/amasad/code/babel/lib/babel/traversal/index.js:62:17)
    at TraversalPath.visit (/Users/amasad/code/babel/lib/babel/traversal/path/index.js:798:31)
    at TraversalContext.visitMultiple (/Users/amasad/code/babel/lib/babel/traversal/context.js:59:16)
    at TraversalContext.visit (/Users/amasad/code/babel/lib/babel/traversal/context.js:80:19)
```