@sebmck 2015-05-07
Grr, looks like it's due to the `es6.arrowFunctions` transformer. Can you try blacklisting it?