@sebmck 2015-05-07
Compiling Traceur which has a ton of ES6 so it's a more reflective benchmark:

`master`

```
real	0m30.877s
user	0m31.675s
sys	0m0.290s
```

`single-pass`

```
real	0m20.517s
user	0m21.308s
sys	0m0.297s
```