@sebmck 2015-05-07
Seeing on average a ~30% improvement on actual ES6 codebases.