@amasad 2015-05-07
A 130-150 seconds build is now a 80-90 seconds.

I wish I had your emoji game right now, but here is my best:

:boom: :facepunch: :stuck_out_tongue_closed_eyes: :dancers: :dancers: