@amasad 2015-05-07
So `shouldVisit` is deprecated in favor of `category`?