@sebmck 2015-05-07
Moreorless, yeah. Doing the prepass itself was pretty expensive, especially on large trees and the "merged" visitors basically eliminates it's usecase.