@amasad 2015-05-07
I don't see any `category` for builtin transforms, are they all in a single pass now? Otherwise on what basis is the merge decided?