@sebmck 2015-05-07
I kinda cheated a bit and add the `category` [here](https://github.com/babel/babel/blob/single-pass/src/babel/transformation/index.js#L12).