@amasad 2015-05-07
I'm surprised they all can now work in a single pass. How did you get rid of ordering?
For example in the comments in  `src/babel/transformation/transformers/index.js` it used to say which transform needs to run first.