@sebmck 2015-05-07
Well they *sorta* work in a single pass. There's still 97/414 tests failing.

Basically, previously the visitor format was:

```javascript
{
  Identifier: {
    enter: Function,
    exit: Function
  }
}
```

Now `enter` and `exit` can be arrays:

```javascript
{
  Identifier: {
    enter: [Function],
    exit: [Function]
  }
}
```

The `mergeStack` method just crawls all the enabled transformers, and when it hits a transformer with a `category` collects all the proceeding transformers with the same category, merges their visitors and creates a new transformer in their place.

Ordering wasn't really removed. If you have two transformers that define `Identifier.enter` methods then the order in which those methods are ran in is determined by the "global position" (ie. just what position the transformer is in that transformers object). It causes some weird issues where enter methods need to be turned to exit ones and that's largely been what most of the changes have been.