@sebmck 2015-05-08
45 failing tests. I split the transformers into 4 different passes, will look into minimising it further in the future but there wasn't really any other way to get around it since you couldn't get around the visitor precedence in some cases.