@amasad 2015-05-08
Ok, can you check in (or create another repo) with your Traceur benchmark? so we can track how we're doing